chore(human-input): merge hitl.py into entities.py
